Add rnp-common object library.
This is intended for code that is shared between the libraries and
executables.
